Improved test task due to comments in pullrequest
